fix map

+++ examples/bot/__main__.py
@@ -34,7 +34,7 @@ def init_config():
     parser.add_argument("-l", "--location", help="Location", required=required("location"))
     parser.add_argument("-k", "--key", help="Google Maps API Key", required=required("key"))
     parser.add_argument("-m", "--minpokemon", type=int, help="Minimum number of pokemon for auto transfing")
-    parser.add_argument("-s", "--speed", type=int, help="Travel speed in miles per hour", required=required("speed"))
+    parser.add_argument("-s", "--speed", type=float, help="Travel speed in miles per hour", required=required("speed"))
     parser.add_argument("-q", "--powerquotient", type=int, help="Minimum power quotient for keeping pokemon")
     parser.add_argument("-d", "--debug", help="Debug Mode", action='store_true')
     parser.add_argument("--best_balls_first", action='store_true', help="Prioritize throwing better balls")
